requirements.txt
absl-py==0.9.0
alabaster==0.7.12
appdirs==1.4.3
argon2-cffi==19.2.0
asgiref==3.2.7
astor==0.8.1
astroid==2.4.1
attrs==19.3.0
autopep8==1.5.2
Babel==2.8.0
backcall==0.2.0
beautifulsoup4==4.9.1
bleach==3.1.5
boto3==1.14.43
botocore==1.17.43
bs4==0.0.1
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
colorama==0.4.3
contextlib2==0.6.0.post1
crayons==0.3.1
cycler==0.10.0
Cython==0.29.16
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.0
dj-database-url==0.5.0
Django==3.0.6
django-appconf==1.0.4
django-braces==1.14.0
django-cors-headers==3.2.1
django-environ==0.4.5
django-extensions==2.2.9
django-filter==2.2.0
django-heroku==0.3.1
django-ipware==2.1.0
django-storages==1.9.1
django-templated-mail==1.1.1
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
docutils==0.15.2
dominate==2.5.2
drf-flex-fields==0.8.0
drf-writable-nested==0.6.1
entrypoints==0.3
filelock==3.0.12
gast==0.4.0
google-pasta==0.2.0
grpcio==1.31.0
gunicorn==20.0.4
idna==2.10
imagecorruptions==1.1.0
imageio==2.9.0
imagesize==1.2.0
importlib-metadata==1.7.0
imutils==0.5.3
isodate==0.6.0
isort==4.3.21
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
jmespath==0.10.0
jsonfield==3.1.0
jsonschema==3.2.0
keyring==21.4.0
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
lxml==4.5.2
Markdown==3.2.2
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
msrest==0.6.14
nbconvert==5.6.1
nbformat==5.0.7
networkx==2.4
nose==1.3.7
notebook==6.1.3
oauthlib==3.1.0
packaging==20.4
parso==0.7.1
pbr==5.4.5
pickleshare==0.7.5
Pillow==7.2.0
pkginfo==1.5.0.1
prometheus-client==0.8.0
prompt-toolkit==3.0.6
protobuf==3.12.4
psycopg2==2.8.5
pycodestyle==2.5.0
pycparser==2.20
Pygments==2.6.1
pylint==2.5.2
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
python-decouple==3.3
pytz==2020.1
PyYAML==5.3.1
pyzmq==19.0.2
readme-renderer==26.0
requests==2.24.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rfc3986==1.4.0
s3transfer==0.3.3
Send2Trash==1.5.0
Shapely==1.7.0
simplejson==3.17.2
six==1.14.0
snowballstemmer==2.0.0
soupsieve==2.0.1
Sphinx==3.2.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sqlparse==0.3.1
stevedore==1.32.0
stripe==2.48.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
tifffile==2020.7.24
toml==0.10.0
tornado==6.0.4
tqdm==4.49.0
traitlets==4.3.3
twine==3.2.0
typed-ast==1.4.1
Unipath==1.1
urllib3==1.25.10
virtualenv==20.0.17
virtualenv-clone==0.5.4
virtualenvwrapper==4.8.4
visitor==0.1.3
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
whitenoise==5.1.0
widgetsnbextension==3.5.1
wrapt==1.12.1
zipp==3.1.0
